Trying to add types and spec to the board
@gabrielelana: what do you think about them?
If the position() type is a good idea maybe it should be good to change
the move signature to pass a position() instead of the Row and the
Column.
I have also updated the check/1 so that we are not tied to 'x' or 'o'
atoms but here I have another doubt. Looking at the symbol() type, the
atom 'none' is a valid symbol() so that I would like to enforce this but
I don't know how in this moment.